Mesures des serveurs en bigint, test d'alerte sans tâche pendante
- server_metrics : mem_bytes, disk_bytes, net_rx et net_tx passent en
  bigint (migration 0040). Au-delà de 2 Gio l'insertion échouait sans
  bruit et le serveur n'avait aucun historique. Test de non-régression
  dans metrics-collector.integration.test.ts.
- Lectures des mesures converties en float8 côté SQL.
- security-alert : l'envoi SMTP simulé est libéré et attendu avant la
  fin du test, pour ne plus interbloquer le truncate du test suivant.

+++ apps/api/src/modules/reseller/reseller-share.service.ts
@@ -154,13 +154,13 @@ export class ResellerShareService {
          * reste lisible et laisse l'index `(server_id, at)` faire son travail.
          */
         memBytes: sql<number | null>`(
-          select m.mem_bytes from ${serverMetrics} m
+          select m.mem_bytes::float8 from ${serverMetrics} m
           where m.server_id = ${servers.id}
             and m.at > now() - interval '${sql.raw(String(MEASUREMENT_MAX_AGE_MS / 1000))} seconds'
           order by m.at desc limit 1
         )`,
         diskBytes: sql<number | null>`(
-          select m.disk_bytes from ${serverMetrics} m
+          select m.disk_bytes::float8 from ${serverMetrics} m
           where m.server_id = ${servers.id}
             and m.at > now() - interval '${sql.raw(String(MEASUREMENT_MAX_AGE_MS / 1000))} seconds'
           order by m.at desc limit 1
